Add session token authentication
This commit also changed the requirements.txt. So, basically as of right
now Starlette's @requires authentication decorator doesn't work on
FastAPI latest. This was fixed in Starlette version 0.13.8, however
FastAPI is on Starlette 0.13.6. There is a PR pending to bump the
version, but it has not been accepting yet for some reason (PR is
approved, but no response from head maintainer after multiple months)

Normally, you should be able simply add the VCS for the PR and call it a
day. But for some reason, that doesn't work here. If you add that to
your requirements.txt, you get an error about how setuptools cannot be
found. This is absolutely nonsensical, because you can literally include
setuptools in your requirements and it will still claim it is not found.

A recurring part of this issue seems to be due to FastAPI using `flit`.
Apparently, some people don't want to use setup.py which is fine, but it
absolutely DESTROYS pip's ability to function correctly. I am not sure
who is to blame here. Anyway, the most confusing part is that installing
from the CLI actually works just fine:

```
pip3 install git+https://github.com/hanneskuettner/fastapi.git@bump-starlette#egg=fastapi
```

However, as soon as you throw that EXACT install into a
requirements.txt, it fails:

```
-e git+https://github.com/hanneskuettner/fastapi.git@bump-starlette#egg=fastapi
```

In this case, I am trying to use hanneskuettner's FastAPI PR, using
branch bump-starlette. `pip` forces you to add `#egg=fastapi` in order
to know what package you're installing because FastAPI doesn't have a
setup.py due to them using flit. However, it doesn't require this on the
CLI, and I have LITERALLY NO IDEA WHY THIS IS THE CASE.

So we are now in a situation where you can install this package USING
PIP JUST FINE, but as soon as you put it into a requirements.txt IT
DOESN'T WORK, FOR NO DISCERNABLE REASON. I'm guessing this has to do
with a poor combination of pip CLI acting different than
requirements.txt, and flit doing whatever it wants to since apparently
they're cool to use setup.py.

Anyways, this is why I have an intentional syntax error in
requirements.txt. It forces a user installing with `-r` to run a command
that pip is capable of running, but for some reason refuses to.