Fix FastAPI requirement
Regarding the last commit, it turns out that this was an issue because
of the -e option. For some reason, the OFFICIAL PYTHON DOCUMENTATION
regarding VCS (https://packaging.python.org/tutorials/installing-packages/#installing-from-vcs)
says to use this -e flag when installing from VCS, with no
explanation as to why. For some reason, this "editable mode" breaks
flit. A shout out for FastAPI for not bumping the Starlette version, to
the Python documentation team for not explaning why you should use this
flag for VCS, and to flit for somehow breaking when this flag is enabled.